requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
blinker==1.8.2
    # via flask
cachetools==5.4.0
    # via google-auth
certifi==2024.7.4
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via flask
flask==3.0.3
    # via -r requirements.in
google-api-core==2.19.1
    # via google-api-python-client
google-api-python-client==2.138.0
    # via -r requirements.in
google-auth==2.32.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.2.0
    # via
    #   -r requirements.in
    #   google-api-python-client
google-auth-oauthlib==1.2.1
    # via -r requirements.in
googleapis-common-protos==1.63.2
    # via google-api-core
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.7
    # via requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via flask
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
oauthlib==3.2.2
    # via requests-oauthlib
proto-plus==1.24.0
    # via google-api-core
protobuf==5.27.2
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via google-auth
pyparsing==3.1.2
    # via httplib2
python-decouple==3.8
    # via -r requirements.in
requests==2.32.3
    # via
    #   google-api-core
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
ruff==0.5.5
    # via -r requirements.in
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.2.2
    # via requests
werkzeug==3.0.3
    # via flask
